Process to Implement Keyword Research Tool
So, I've created a Keyword Research Tool with a adword account. For This I needed to read the Google ads Documentation(https://developers.google.com/google-ads/api/docs/start) and also to create a google-ads test account. Basically the suggestion for the viewers that go through the document, understand it properly and then run the code, in this way you can modify the code as your experties. Thank You!